Refactor RSS2
